+++ src/view/game_view_trait.rs
@@ -1,6 +1,6 @@
 use crate::game::game_state::GameState;
 
 pub trait GameViewTrait {
-    fn new(game_state: GameState, cell_size: f64, update_interval_ms: u64) -> Self where Self: Sized;
+    fn new(game_state: GameState, cell_size: usize, update_interval_ms: usize) -> Self where Self: Sized;
     fn init(&mut self) -> Result<(), String>;
 }
